Update to jquery 1.10.2
- index.html & upload.html now working with browserify change

+++ build.sh
@@ -1,15 +1,18 @@
 #!/bin/sh
 
 set -e
+debug=1
 
-rm -rf build
-mkdir build
+mkdir -p build
+rm -rf build/*
 
-#cp -r css/images build
-#cp index.html build
+cp -r app/html/ build
+cp -r app/images build
 
 #echo "Combining css and minifying..."
 #cat css/bootstrap-tour.min.css css/dge.css css/venn.css css/slick.grid.css | cleancss > build/main.min.css
+cp -r app/css build
+
 
 echo "Compiling CoffeeScript and bundling all js..."
 for f in app/js/*-req.coffee; do
@@ -17,7 +20,12 @@ for f in app/js/*-req.coffee; do
   b=${t/.js/-big.js}
   echo "Building $f -> $t"
 
-  browserify -t coffeeify -t hbsfy $f > $b
-  uglifyjs $b > $t
-  rm $b
+  if [ $debug ]; then
+      browserify --debug -t coffeeify -t hbsfy $f > $t
+  else
+      browserify -t coffeeify -t hbsfy $f > $b
+      uglifyjs $b > $t
+      cp $b > $t
+      rm $b
+  fi
 done
